requirements.txt
# Core dependencies
chromadb==0.6.3
streamlit==1.41.1
ollama==0.6.1

# LlamaIndex and related packages
llama-index-core==0.12.11
llama-index-embeddings-ollama==0.5.0
llama-index-llms-ollama==0.5.0
llama-index-readers-file==0.4.3
llama-index-vector-stores-chroma==0.4.1

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1

# Utilities
requests==2.32.0
python-dotenv==1.0.1
